Adress Sessions and Documentation Feedback
