+++ Sources/StytchUI/Documentation.docc/Documentation.md
@@ -7,6 +7,8 @@ Provides a pre-built, customizable, themeable, dark-and-light mode capable UI fo
 
 The `StytchUI` client contains our pre-built UI components, and incorporates our headless SDK under-the-hood, so you can use both together, or the UI on it's own, depending on your project's needs.
 
+[Example code for StytchUI here!](https://github.com/stytchauth/stytch-ios/blob/main/tutorials/UI.md)
+
 Currently, only Stytch Consumer projects are supported.
 
 The following products are fully supported:
